+++ core/src/test/scala/org/apache/spark/sql/delta/DeltaCDCSuite.scala
@@ -38,6 +38,8 @@ abstract class DeltaCDCSuiteBase
   with SharedSparkSession  with CheckCDCAnswer
   with DeltaSQLCommandTest {
 
+  import testImplicits._
+
   override protected def sparkConf: SparkConf = super.sparkConf
     .set(DeltaConfigs.CHANGE_DATA_FEED.defaultTablePropertyKey, "true")
 
@@ -627,6 +629,62 @@ abstract class DeltaCDCSuiteBase
           spark.range(30)
             .withColumn("_change_type", lit("insert"))
             .withColumn("_commit_version", (col("id") / 10).cast(LongType)))
+      }
+    }
+  }
+
+  test("batch write: append, dynamic partition overwrite + CDF") {
+    withSQLConf(
+      DeltaConfigs.CHANGE_DATA_FEED.defaultTablePropertyKey -> "true",
+      DeltaSQLConf.DYNAMIC_PARTITION_OVERWRITE_ENABLED.key -> "true") {
+      withTempDir { tempDir =>
+        def data: DataFrame = spark.read.format("delta").load(tempDir.toString)
+
+        Seq(("a", "x"), ("b", "y"), ("c", "x")).toDF("value", "part")
+          .write
+          .format("delta")
+          .partitionBy("part")
+          .mode("append")
+          .save(tempDir.getCanonicalPath)
+        checkAnswer(
+          cdcRead(new TablePath(tempDir.getCanonicalPath), StartingVersion("0"), EndingVersion("0"))
+            .drop(CDC_COMMIT_TIMESTAMP),
+          Row("a", "x", "insert", 0) :: Row("b", "y", "insert", 0) ::
+            Row("c", "x", "insert", 0) :: Nil
+        )
+
+        // ovewrite nothing
+        Seq(("d", "z")).toDF("value", "part")
+          .write
+          .format("delta")
+          .partitionBy("part")
+          .mode("overwrite")
+          .option(DeltaOptions.PARTITION_OVERWRITE_MODE_OPTION, "dynamic")
+          .save(tempDir.getCanonicalPath)
+        checkDatasetUnorderly(data.select("value", "part").as[(String, String)],
+          ("a", "x"), ("b", "y"), ("c", "x"), ("d", "z"))
+        checkAnswer(
+          cdcRead(new TablePath(tempDir.getCanonicalPath), StartingVersion("1"), EndingVersion("1"))
+            .drop(CDC_COMMIT_TIMESTAMP),
+          Row("d", "z", "insert", 1) :: Nil
+        )
+
+        // overwrite partition `part`="x"
+        Seq(("a", "x"), ("e", "x")).toDF("value", "part")
+          .write
+          .format("delta")
+          .partitionBy("part")
+          .mode("overwrite")
+          .option(DeltaOptions.PARTITION_OVERWRITE_MODE_OPTION, "dynamic")
+          .save(tempDir.getCanonicalPath)
+        checkDatasetUnorderly(data.select("value", "part").as[(String, String)],
+          ("a", "x"), ("b", "y"), ("d", "z"), ("e", "x"))
+        checkAnswer(
+          cdcRead(new TablePath(tempDir.getCanonicalPath), StartingVersion("2"), EndingVersion("2"))
+            .drop(CDC_COMMIT_TIMESTAMP),
+          Row("a", "x", "delete", 2) :: Row("c", "x", "delete", 2) ::
+            Row("a", "x", "insert", 2) :: Row("e", "x", "insert", 2) :: Nil
+        )
       }
     }
   }
